@@ -3,41 +3,43 @@
   "description": "web ui for interacting with opentutor",
   "version": "1.0.0",
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@reach/router": "^1.3.4",
-    "@xapi/cmi5": "^1.1.1",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@iostindex/gatsby-plugin-material-ui": "^5.0.0",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/material": "^5.12.1",
+    "@xapi/cmi5": "^1.1.2",
     "axios": "^0.21.1",
-    "gatsby": "^3.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "gatsby": "^5.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-player": "^2.9.0",
     "react-scroll": "^1.8.2",
-    "reading-time": "^1.3.0",
-    "styled-components": "^5.3.0"
+    "reading-time": "^1.5.0",
+    "stream-browserify": "^3.0.0",
+    "styled-components": "^5.3.0",
+    "tss-react": "^4.8.2",
+    "util": "^0.12.5"
   },
   "devDependencies": {
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "@types/react-scroll": "^1.8.2",
-    "@typescript-eslint/eslint-plugin": "^4.27.0",
-    "@typescript-eslint/parser": "^4.27.0",
-    "eslint": "^7.28.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-loader": "^3.0.4",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "gatsby-cli": "^3.7.1",
-    "gatsby-plugin-env-variables": "^2.1.0",
-    "gatsby-plugin-eslint": "^3.0.0",
-    "gatsby-plugin-material-ui": "^3.0.1",
-    "gatsby-plugin-styled-components": "^4.7.1",
-    "gatsby-plugin-typescript": "^3.7.1",
+    "@types/react": "^18.0.38",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-scroll": "^1.8.6",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-react-app": "^7.0.1",
+    "gatsby-plugin-env-variables": "^2.3.0",
+    "gatsby-plugin-eslint": "^4.0.4",
+    "gatsby-plugin-manifest": "^5.5.0",
+    "gatsby-plugin-sitemap": "^6.5.0",
+    "gatsby-plugin-styled-components": "^6.9.0",
+    "gatsby-plugin-typescript": "^5.5.0",
     "license-check-and-add": "^4.0.2",
     "npm-check-updates": "^11.6.0",
-    "typescript": "^4.3.2"
+    "react-markdown": "^5.0.3",
+    "typescript": "^4.5.0"
   },
   "keywords": [
     "gatsby"
