@@ -1,7 +1,11 @@
 import React from "react";
-import { createMuiTheme, MuiThemeProvider } from "@material-ui/core/styles";
+import {
+  createTheme,
+  ThemeProvider,
+  StyledEngineProvider,
+} from "@mui/material/styles";
 
-const theme = createMuiTheme({
+const theme = createTheme({
   palette: {
     primary: {
       main: "#1b6a9c",
@@ -11,5 +15,7 @@ const theme = createMuiTheme({
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export const wrapRootElement = ({ element }) => (
-  <MuiThemeProvider theme={theme}>{element}</MuiThemeProvider>
+  <StyledEngineProvider injectFirst>
+    <ThemeProvider theme={theme}>{element}</ThemeProvider>
+  </StyledEngineProvider>
 );
