@@ -7,87 +7,21 @@ The full terms of this copyright and license should always be found in the root 
 import React, { useRef } from "react";
 import ReactPlayer from "react-player/youtube";
 import clsx from "clsx";
-import ZoomInIcon from "@material-ui/icons/ZoomIn";
-import ReplayIcon from "@material-ui/icons/Replay";
-import { makeStyles } from "@material-ui/core/styles";
-import { IconButton } from "@material-ui/core";
-import { Typography } from "@material-ui/core";
+import ZoomInIcon from "@mui/icons-material/ZoomIn";
+import ReplayIcon from "@mui/icons-material/Replay";
+import { makeStyles } from "tss-react/mui";
+import { IconButton } from "@mui/material";
+import { Typography } from "@mui/material";
 import { fetchLesson } from "api";
 import { Lesson, LessonFormat, Media, MediaType } from "types";
 import withLocation from "wrap-with-location";
 import ImageDialog from "./ImageDialog";
 
-const useStyles = makeStyles((theme) => ({
-  scroll: {
-    backgroundColor: theme.palette.primary.dark,
-    position: "relative",
-    top: 0,
-    left: 0,
-    width: "100%",
-    height: "100%",
-    overflow: "auto",
-    whiteSpace: "nowrap",
-  },
-  image: {
-    position: "relative",
-    top: 0,
-    left: 0,
-    minWidth: 400,
-  },
-  innerZoomOverlay: {
-    position: "absolute",
-    zIndex: 1,
-    bottom: 0,
-    right: 0,
-    height: 50,
-    width: 50,
-    color: "white",
-  },
-  centerLock: {
-    position: "absolute",
-    top: "50%",
-    left: "calc(50% + 0px)",
-    transform: "translate(-50%, -50%)",
-  },
-  released: {
-    marginTop: "-30",
-    transform: "translate(0%, -60%)",
-    padding: 10,
-    height: 10,
-    width: "140%",
-  },
-  censored: {
-    borderRadius: 10,
-    background: theme.palette.primary.main,
-    color: "white",
-    padding: 10,
-    height: 10,
-    width: "calc(100% - 20px)",
-    position: "relative",
-    borderColor: "orange",
-    borderWidth: 2,
-    borderStyle: "solid",
-  },
-  survey: {
-    padding: theme.spacing(2),
-    maxWidth: 500,
-    marginLeft: "auto",
-    marginRight: "auto",
-  },
-  mediaRoot: {},
-  mediaDefault: {
-    height: "35%",
-  },
-  mediaSurveySays: {
-    height: "30%",
-  },
-}));
-
 const LessonMedia = (props: {
   search: { lesson: string };
   lessonFormat: string;
 }): JSX.Element => {
-  const styles = useStyles();
+  const { classes: styles } = useStyles();
   const { lesson } = props.search;
   const [media, setMedia] = React.useState<Media | undefined>(undefined);
   const [isVideoOver, setIsVideoOver] = React.useState(false);
@@ -265,5 +199,71 @@ const LessonMedia = (props: {
     return <></>;
   }
 };
+
+const useStyles = makeStyles({ name: { LessonMedia } })((theme) => ({
+  scroll: {
+    backgroundColor: theme.palette.primary.dark,
+    position: "relative",
+    top: 0,
+    left: 0,
+    width: "100%",
+    height: "100%",
+    overflow: "auto",
+    whiteSpace: "nowrap",
+  },
+  image: {
+    position: "relative",
+    top: 0,
+    left: 0,
+    minWidth: 400,
+  },
+  innerZoomOverlay: {
+    position: "absolute",
+    zIndex: 1,
+    bottom: 0,
+    right: 0,
+    height: 50,
+    width: 50,
+    color: "white",
+  },
+  centerLock: {
+    position: "absolute",
+    top: "50%",
+    left: "calc(50% + 0px)",
+    transform: "translate(-50%, -50%)",
+  },
+  released: {
+    marginTop: "-30",
+    transform: "translate(0%, -60%)",
+    padding: 10,
+    height: 10,
+    width: "140%",
+  },
+  censored: {
+    borderRadius: 10,
+    background: theme.palette.primary.main,
+    color: "white",
+    padding: 10,
+    height: 10,
+    width: "calc(100% - 20px)",
+    position: "relative",
+    borderColor: "orange",
+    borderWidth: 2,
+    borderStyle: "solid",
+  },
+  survey: {
+    padding: theme.spacing(2),
+    maxWidth: 500,
+    marginLeft: "auto",
+    marginRight: "auto",
+  },
+  mediaRoot: {},
+  mediaDefault: {
+    height: "35%",
+  },
+  mediaSurveySays: {
+    height: "30%",
+  },
+}));
 
 export default withLocation(LessonMedia);
