@@ -5,75 +5,31 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 import React, { useState } from "react";
-import {
-  createStyles,
-  Theme,
-  useTheme,
-  withStyles,
-  WithStyles,
-} from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
-import Dialog from "@material-ui/core/Dialog";
-import MuiDialogTitle from "@material-ui/core/DialogTitle";
-import MuiDialogContent from "@material-ui/core/DialogContent";
-import MuiDialogActions from "@material-ui/core/DialogActions";
-import IconButton from "@material-ui/core/IconButton";
-import CloseIcon from "@material-ui/icons/Close";
-import Typography from "@material-ui/core/Typography";
+import { useTheme, styled } from "@mui/material/styles";
+import Button from "@mui/material/Button";
+import Dialog from "@mui/material/Dialog";
+import MuiDialogTitle from "@mui/material/DialogTitle";
+import MuiDialogContent from "@mui/material/DialogContent";
+import MuiDialogActions from "@mui/material/DialogActions";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
+import Typography from "@mui/material/Typography";
 import { SummaryIndicator } from "components/TargetIndicator";
 import { Target } from "types";
-import { useMediaQuery } from "@material-ui/core";
+import { useMediaQuery } from "@mui/material";
 
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      margin: 0,
-      padding: theme.spacing(2),
-    },
-  });
-
-export interface DialogTitleProps extends WithStyles<typeof styles> {
-  id: string;
-  children: React.ReactNode;
-  onClose: () => void;
-}
-
-const DialogTitle = withStyles(styles)((props: DialogTitleProps) => {
-  const { children, classes, onClose, ...other } = props;
-  return (
-    <MuiDialogTitle disableTypography className={classes.root} {...other}>
-      <Typography variant="h6">{children}</Typography>
-      {onClose ? (
-        <IconButton
-          aria-label="close"
-          style={{
-            position: "absolute",
-            right: 20,
-            top: 20,
-            color: "gray",
-          }}
-          onClick={onClose}
-        >
-          <CloseIcon />
-        </IconButton>
-      ) : null}
-    </MuiDialogTitle>
-  );
-});
-
-const DialogContent = withStyles((theme: Theme) => ({
+const DialogContent = styled(MuiDialogContent)(({ theme }) => ({
   root: {
     padding: theme.spacing(2),
   },
-}))(MuiDialogContent);
+}));
 
-const DialogActions = withStyles((theme: Theme) => ({
+const DialogActions = styled(MuiDialogActions)(({ theme }) => ({
   root: {
     margin: 0,
     padding: theme.spacing(1),
   },
-}))(MuiDialogActions);
+}));
 
 export default function SummaryPopup(props: {
   open: boolean;
@@ -85,25 +41,42 @@ export default function SummaryPopup(props: {
   const { open, onCloseRequested, buttonText, targets } = props;
   const [tranState, setTranState] = useState("summary-popup-trans-none");
   const theme = useTheme();
-  const fullScreen = useMediaQuery(theme.breakpoints.down("xs"));
+  const fullScreen = useMediaQuery(theme.breakpoints.down("sm"));
 
   return (
     <Dialog
       data-cy="summary-popup"
       onClose={onCloseRequested}
-      onEntered={() => setTranState("summary-popup-trans-done")}
+      TransitionProps={{
+        onEntered: () => setTranState("summary-popup-trans-done"),
+      }}
       aria-labelledby="customized-dialog-title"
       open={open}
       fullScreen={fullScreen}
     >
       <div data-cy={tranState} />
-      <DialogTitle
+
+      <MuiDialogTitle
         id="customized-dialog-title"
         data-cy="customized-dialog-title"
-        onClose={onCloseRequested}
       >
-        Lesson Summary
-      </DialogTitle>
+        <Typography variant="h6">Lesson Summary</Typography>
+        {onCloseRequested ? (
+          <IconButton
+            aria-label="close"
+            style={{
+              position: "absolute",
+              right: 20,
+              top: 20,
+              color: "gray",
+            }}
+            onClick={onCloseRequested}
+          >
+            <CloseIcon />
+          </IconButton>
+        ) : null}
+      </MuiDialogTitle>
+
       <DialogContent dividers>
         <SummaryIndicator targets={targets} />
       </DialogContent>
