@@ -5,6 +5,7 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 import {
+  cyMockDefault,
   cyMockDialog,
   cyMockSession,
   cySetup,
@@ -44,7 +45,7 @@ describe("screenshots - chat responses", () => {
     },
   ].forEach((x, i) => {
     it(`sends message and displays response for lesson ${x.lesson}`, () => {
-      cySetup(cy);
+      cyMockDefault(cy);
       cyMockDialog(cy, x.lesson, x.fixtureLessonStart).as("start");
       cyMockSession(cy, x.lesson, x.fixtureLessonContinue).as("response");
       cyVisitWithTestingParam(cy, `/?lesson=${x.lesson}&guest=guest`);
