@@ -5,6 +5,7 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 import {
+  cyMockDefault,
   cyMockDialog,
   cyMockSession,
   cySetup,
@@ -13,7 +14,7 @@ import {
 
 describe("Expectation summary pop-up", () => {
   it("opens with View Summary button", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
     cy.get("[data-cy=target-0-0]").trigger("mouseover").click();
@@ -33,7 +34,7 @@ describe("Expectation summary pop-up", () => {
     },
   ].forEach((x) => {
     it(`shows an icon for each expectation for lesson with ${x.expectedExpectations}`, () => {
-      cySetup(cy);
+      cyMockDefault(cy);
       cyMockDialog(cy, x.lesson, x.fixture);
       cyVisitWithTestingParam(cy, `/?lesson=${x.lesson}&guest=guest`);
       cy.get("[data-cy=target-0-0]").trigger("mouseover").click();
@@ -45,7 +46,7 @@ describe("Expectation summary pop-up", () => {
   });
 
   it("displays on tap any expectation-progress indicator", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
     cy.fixture("q2-1-p1.json").then((desiredServerResponse) => {
@@ -62,10 +63,15 @@ describe("Expectation summary pop-up", () => {
   });
 
   it("displays when lesson is completed", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyMockSession(cy, "q2", "q2-1-p2.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
+    cy.wait(4000);
+    cy.get("[data-cy=chat-msg-1]").should(
+      "have.text",
+      "With a DC input source, does current flow in the same or the opposite direction of the diode arrow?"
+    );
     cy.get("[data-cy=outlined-multiline-static]").type("fake answer");
     cy.get("[data-cy=submit-button]", { timeout: 15000 })
       .should("not.be.disabled")
@@ -75,7 +81,7 @@ describe("Expectation summary pop-up", () => {
   });
 
   it("hides text for expectations that have not been completed", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
     cy.get("[data-cy=target-0-0]").trigger("mouseover").click();
@@ -85,7 +91,7 @@ describe("Expectation summary pop-up", () => {
   });
 
   it("shows no progress for expectations that have not been completed", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
     cy.get("[data-cy=target-0-0]").trigger("mouseover").click();
@@ -95,10 +101,15 @@ describe("Expectation summary pop-up", () => {
   });
 
   it("shows text for expectations that have been completed", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyMockSession(cy, "q2", "q2-1-p2.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
+    cy.wait(4000);
+    cy.get("[data-cy=chat-msg-1]").should(
+      "have.text",
+      "With a DC input source, does current flow in the same or the opposite direction of the diode arrow?"
+    );
     cy.get("[data-cy=outlined-multiline-static]").type("fake answer");
     cy.get("[data-cy=submit-button]").should("be.visible");
     cy.get("[data-cy=submit-button]", { timeout: 15000 })
@@ -114,10 +125,15 @@ describe("Expectation summary pop-up", () => {
   });
 
   it("shows progress for expectations that have been completed", () => {
-    cySetup(cy);
+    cyMockDefault(cy);
     cyMockDialog(cy, "q2", "q2-1-p1.json");
     cyMockSession(cy, "q2", "q2-1-p2.json");
     cyVisitWithTestingParam(cy, "/?lesson=q2&guest=guest");
+    cy.wait(4000);
+    cy.get("[data-cy=chat-msg-1]").should(
+      "have.text",
+      "With a DC input source, does current flow in the same or the opposite direction of the diode arrow?"
+    );
     cy.get("[data-cy=outlined-multiline-static]").type("fake answer");
     cy.get("[data-cy=submit-button]").should("be.visible");
     cy.get("[data-cy=submit-button]", { timeout: 15000 })
